More comprehensive route override  mapping
This fixes an issue where setting a route via the override flag
to a hostname with a protocol would fall through back to default
config routing. It is fixed with more robust validating and
setting of the host and/or port.

However this does introduce a requirement that route overrides
using a hostname MUST set a protocol, which is good practice.
This is due to how Node.js url.parse() works.

 - new route mapping function and tests
 - adjustments to the router to adopt new function
 - use Node.js url.parse() to parse a url
 - validate IPv4 routes
 - fix bug where a valid ip address of 7 characters (8.8.8.8)
   would be ignored
 - update README to reflect changes and add example of how route
   override might be useful