Change test commands and add snyk test to build
This changes the makefile, travis config and npm scripts in order to follow some common conventions.

This allows now for us to run snyk test as part of the build process in travis.

Also adds .snyk file with a policy to ignore the only present vulnerability for 30 days - this allows all tests to pass successfully.