Update travis to allow failure on unsupported node
As Node 9 has just been released, this change adds tests for node versions 8 and 9 but allows tests for version 9 to fail, as it's technically not supported by this version of shunter.

It also removes the old GCC 4.8 config, which shouldn't be needed anymore now that Travis has migrated to Ubuntu Trusty.