Replace sinon.reset() with resetHistory()
sinon.reset() is deprecated so after we upgraded from sinon 2.x to 4.x in #256 we started getting deprecation warnings when running our tests:

```
sinon.reset is deprecated and will be removed from the public API in a future version of sinon.
```

This replaces the calls to reset() with resetHistory() which has exactly the same functionality and parameters.

Addresses #257.