Fix test that was causing child tests to fail
In sinon 4.x, `reset()` has been deprecated for spies and replaced with `resetHistory()`.

As part of `534d1bd` a call to `stub.reset()` was wrongly replaced with `stub.resetHistory()`. `reset()` is not deprecated for stubs, only for spies, so we don't want to change this.

This was causing the client tests in some child apps to fail.