Remove deprecated Buffer() constructor (#271)
This removes all the calls to the deprecated `Buffer()` constructor and replaces them with `Buffer.from()` (for strings) and `Buffer.allocUnsafe()` (for numbers). These are currently used in tests only, so using `allocUnsafe()` doesn't pose a security risk.

The `Buffer()` function and `new Buffer()` constructor have been deprecated since node 6 due to API usability issues that can potentially lead to accidental security issues. As of v10.0.0, a deprecation warning is printed at runtime when `--pending-deprecation` is used or when the calling code is outside node_modules.

Ref: https://nodejs.org/api/deprecations.html#deprecations_dep0005_buffer_constructor