Remove shunter-test-client tool
The shunter-test-client tool that shunter provides in order to test client side JS hasn't been updated in a very long time. No one seems to be using this feature so we rather remove it so it's easier to maintain.

This PR removes the tool, any required dependencies, plus any references in the documentation and config files.